feat(raspberrypi): better raspberry pi support
This primarily was a set of changes so I could have better support for
deploying atuin on Raspberry Pi devices. Basically, I pre-build
binaries for the Pi and host them myself.

This also adds in support for fixing the hostname of atuin, which is
needed because sometimes things like Tailscale muck it up. Right now
you can define this on a one-off using `atuin.hostname` or it will
take the `inventory_hostname` variable (which is what you call the
node in ansible) if not defined. This is used becuase sometimes the
fqdn gets mucked up on these systems for the same reasons that atuin
gets the wrong hostname.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>