fix(sync): fix sync server address
The search and replace for the sync server wasn't handling the situation
where the line with the sync_server information could be commented. This
caused the sync_server to be put at the end of the file, which would put
it in a different section of the TOML file, and therefore, fail.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>