fix(install): don't try to install from script on MacOS
I'm not sure how this happened, but it appears there are times when
it might try to install from a script on MacOS because it wasn't
checking to make sure that it needed to download the program. I added
the extra `ansible_fact` check to be double sure.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>